Refactor poller: Decouple Polling from Decision Task Execution (#189)
