+++ src/main/java/com/uber/cadence/internal/worker/Poller.java
@@ -17,100 +17,61 @@
 
 package com.uber.cadence.internal.worker;
 
+import com.google.common.base.Preconditions;
 import com.uber.cadence.internal.common.BackoffThrottler;
 import com.uber.cadence.internal.metrics.MetricsType;
 import com.uber.m3.tally.Scope;
-import java.util.concurrent.ArrayBlockingQueue;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.ThreadPoolExecutor;
-import java.util.concurrent.TimeUnit;
+
+import java.util.Objects;
+import java.util.concurrent.*;
 import java.util.concurrent.atomic.AtomicReference;
+import java.util.function.Consumer;
+import org.apache.thrift.TException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-/** Calls a passed task in a loop according to {@link PollerOptions}. */
-final class Poller implements SuspendableWorker {
+public final class Poller<T> implements SuspendableWorker {
+
+  interface PollTask<TT> {
+    TT poll() throws TException;
+  }
 
   interface ThrowingRunnable {
     void run() throws Throwable;
   }
 
-  private class PollServiceTask implements Runnable {
-
-    private final ThrowingRunnable task;
-
-    PollServiceTask(ThrowingRunnable task) {
-      this.task = task;
-    }
-
-    @Override
-    public void run() {
-      try {
-        if (pollExecutor.isTerminating()) {
-          return;
-        }
-        pollBackoffThrottler.throttle();
-        if (pollExecutor.isTerminating()) {
-          return;
-        }
-        if (pollRateThrottler != null) {
-          pollRateThrottler.throttle();
-        }
-
-        CountDownLatch suspender = Poller.this.suspendLatch.get();
-        if (suspender != null) {
-          if (log.isDebugEnabled()) {
-            log.debug("poll task suspending latchCount=" + suspender.getCount());
-          }
-          suspender.await();
-        }
-
-        if (pollExecutor.isTerminating()) {
-          return;
-        }
-        task.run();
-        pollBackoffThrottler.success();
-      } catch (Throwable e) {
-        pollBackoffThrottler.failure();
-        if (!(e.getCause() instanceof InterruptedException)) {
-          uncaughtExceptionHandler.uncaughtException(Thread.currentThread(), e);
-        }
-      } finally {
-        // Resubmit itself back to pollExecutor
-        if (!pollExecutor.isTerminating()) {
-          pollExecutor.execute(this);
-        } else {
-          log.info("poll loop done");
-        }
-      }
-    }
-  }
-
-  private static final Logger log = LoggerFactory.getLogger(Poller.class);
-
-  private final PollerOptions options;
-
   private final String identity;
-
-  private final ThrowingRunnable task;
-
+  private final Consumer<T> consumer;
+  private final PollTask<T> pollTask;
+  private final PollerOptions pollerOptions;
+  private static final Logger log = LoggerFactory.getLogger(Poller.class);
   private ThreadPoolExecutor pollExecutor;
+  private final Scope metricsScope;
 
   private final AtomicReference<CountDownLatch> suspendLatch = new AtomicReference<>();
 
   private BackoffThrottler pollBackoffThrottler;
-
   private Throttler pollRateThrottler;
-
-  private final Scope metricsScope;
 
   private Thread.UncaughtExceptionHandler uncaughtExceptionHandler =
       (t, e) -> log.error("Failure in thread " + t.getName(), e);
 
-  Poller(PollerOptions options, String identity, ThrowingRunnable task, Scope metricsScope) {
-    this.options = options;
+  Poller(
+      String identity,
+      PollTask<T> pollTask,
+      Consumer<T> consumer,
+      PollerOptions pollerOptions,
+      Scope metricsScope) {
+    Objects.requireNonNull(identity, "identity cannot be null");
+    Objects.requireNonNull(pollTask, "poll service should not be null");
+    Objects.requireNonNull(consumer, "consumer should not be null");
+    Objects.requireNonNull(pollerOptions, "pollerOptions should not be null");
+    Objects.requireNonNull(metricsScope, "metricsScope should not be null");
+
     this.identity = identity;
-    this.task = task;
+    this.pollTask = pollTask;
+    this.consumer = consumer;
+    this.pollerOptions = pollerOptions;
     this.metricsScope = metricsScope;
   }
 
@@ -119,12 +80,12 @@ final class Poller implements SuspendableWorker {
     if (log.isInfoEnabled()) {
       log.info("start(): " + toString());
     }
-    if (options.getMaximumPollRatePerSecond() > 0.0) {
+    if (pollerOptions.getMaximumPollRatePerSecond() > 0.0) {
       pollRateThrottler =
           new Throttler(
               "poller",
-              options.getMaximumPollRatePerSecond(),
-              options.getMaximumPollRateIntervalMilliseconds());
+              pollerOptions.getMaximumPollRatePerSecond(),
+              pollerOptions.getMaximumPollRateIntervalMilliseconds());
     }
 
     // It is important to pass blocking queue of at least options.getPollThreadCount() capacity.
@@ -132,22 +93,22 @@ final class Poller implements SuspendableWorker {
     // releases a thread.
     pollExecutor =
         new ThreadPoolExecutor(
-            options.getPollThreadCount(),
-            options.getPollThreadCount(),
+            pollerOptions.getPollThreadCount(),
+            pollerOptions.getPollThreadCount(),
             1,
             TimeUnit.SECONDS,
-            new ArrayBlockingQueue<>(options.getPollThreadCount()));
+            new ArrayBlockingQueue<>(pollerOptions.getPollThreadCount()));
     pollExecutor.setThreadFactory(
         new ExecutorThreadFactory(
-            options.getPollThreadNamePrefix(), options.getUncaughtExceptionHandler()));
+            pollerOptions.getPollThreadNamePrefix(), pollerOptions.getUncaughtExceptionHandler()));
 
     pollBackoffThrottler =
         new BackoffThrottler(
-            options.getPollBackoffInitialInterval(),
-            options.getPollBackoffMaximumInterval(),
-            options.getPollBackoffCoefficient());
-    for (int i = 0; i < options.getPollThreadCount(); i++) {
-      pollExecutor.execute(new PollServiceTask(task));
+            pollerOptions.getPollBackoffInitialInterval(),
+            pollerOptions.getPollBackoffMaximumInterval(),
+            pollerOptions.getPollBackoffCoefficient());
+    for (int i = 0; i < pollerOptions.getPollThreadCount(); i++) {
+      pollExecutor.execute(new PollLoopTask(new PollExecutionTask()));
       metricsScope.counter(MetricsType.POLLER_START_COUNTER).inc(1);
     }
   }
@@ -167,7 +128,7 @@ final class Poller implements SuspendableWorker {
 
   @Override
   public void shutdownNow() {
-    log.info("shutdownNow poller=" + this.options.getPollThreadNamePrefix());
+    log.info("shutdownNow poller=" + this.pollerOptions.getPollThreadNamePrefix());
     if (!isStarted()) {
       return;
     }
@@ -215,6 +176,79 @@ final class Poller implements SuspendableWorker {
 
   @Override
   public String toString() {
-    return "Poller{" + "options=" + options + ", identity=" + identity + '}';
+    return "Poller{" + "options=" + pollerOptions + ", identity=" + identity + '}';
+  }
+
+  private class PollLoopTask implements Runnable {
+
+    private final Poller.ThrowingRunnable task;
+
+    PollLoopTask(Poller.ThrowingRunnable task) {
+      this.task = task;
+    }
+
+    @Override
+    public void run() {
+      try {
+        if (pollExecutor.isTerminating()) {
+          return;
+        }
+        pollBackoffThrottler.throttle();
+        if (pollExecutor.isTerminating()) {
+          return;
+        }
+        if (pollRateThrottler != null) {
+          pollRateThrottler.throttle();
+        }
+
+        CountDownLatch suspender = Poller.this.suspendLatch.get();
+        if (suspender != null) {
+          if (log.isDebugEnabled()) {
+            log.debug("poll task suspending latchCount=" + suspender.getCount());
+          }
+          suspender.await();
+        }
+
+        if (pollExecutor.isTerminating()) {
+          return;
+        }
+        task.run();
+        pollBackoffThrottler.success();
+      } catch (Throwable e) {
+        pollBackoffThrottler.failure();
+        if (!(e.getCause() instanceof InterruptedException)) {
+          uncaughtExceptionHandler.uncaughtException(Thread.currentThread(), e);
+        }
+      } finally {
+        // Resubmit itself back to pollExecutor
+        if (!pollExecutor.isTerminating()) {
+          pollExecutor.execute(this);
+        } else {
+          log.info("poll loop done");
+        }
+      }
+    }
+  }
+
+  private class PollExecutionTask implements Poller.ThrowingRunnable {
+    private Semaphore pollSemaphore;
+
+    PollExecutionTask() {
+      this.pollSemaphore = new Semaphore(pollerOptions.getPollThreadCount());
+    }
+
+    @Override
+    public void run() throws Exception {
+      try {
+        pollSemaphore.acquire();
+        T task = pollTask.poll();
+        if (task == null) {
+          return;
+        }
+        consumer.accept(task);
+      } finally {
+        pollSemaphore.release();
+      }
+    }
   }
 }
